Revert "Add a binder badge"
This reverts commit f6720a3777bfacdaa81f241341885c0a09d90150.

While it's good to be able to run the latest stable-4.11 master of
GAP on Binder, connecting this repisitory to Binder will have
little practical use. Instead of starting the latest nightly build
of the Docker container for the stable-4.11 branch from DockerHub,
if there are any newer commits to this branch, it will use the
Dockerfile from this repository to run the whole build of GAP and
packages on Binder. That will take too long. Instead of that, one
should have another repository, looking like e.g.
https://github.com/gap-system/gap-docker-binder/, which will always
take the latest container from DockerHub and launch it.